* Compiled using VS 2008

+++ guids.h
@@ -34,7 +34,7 @@ DEFINE_GUID(MEDIASUBTYPE_AVI_DTS,
 DECLARE_INTERFACE_(IAC3File, IUnknown)
 {
   // Stats
-  STDMETHOD (get_info)  (char *info, int len) = 0;
+  STDMETHOD (get_info)  (char *info, size_t len) = 0;
   STDMETHOD (get_pos)   (unsigned *frames, unsigned *bytes, unsigned *ms) = 0;
   STDMETHOD (get_size)  (unsigned *frames, unsigned *bytes, unsigned *ms) = 0;
 };
